+++ frontend/lost/src/containers/Annotation/SIA/lost-sia/CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased] - 
+### Added
+- Delete annotation in creation mode when hitting *Escape*-Key
+### Changed
+- Be able to deal with mixed color possible labels -> where a part of labels 
+  will have a specified color and the other part has no provided color
+
 ## [0.8.0] - 2021-10-14
 ### Added
 - Added lockedAnnos prop for Canvas in order to lock annos by id. Locked annos